refactor(qwen3coder): split call-parsing primitives into a companion module
qwen3coder-protocol.ts had grown past 3,100 lines. The parsing primitives
(tag scanning, parameter-tag parsing, malformed-opener normalization,
tool-name salvage, schema-property maps) move to
qwen3coder-call-parsing.ts, following the existing companion-module
pattern (qwen3coder-stream-call-content.ts). The protocol file keeps the
prompt formatting, generate-path assembly, and streaming state machine.
No behavior change; the tool-name salvage test now imports from the new
module.
